Added everything to rerun feedback and work with feedback model versions

+++ WriterDesk1/backend/app/models.py
@@ -172,7 +172,7 @@ class Scores(db.Model):
     scoreStructure   = db.Column(db.Numeric(4,2), unique=False, default=None)
     scoreIntegration = db.Column(db.Numeric(4,2), unique=False, default=None)
     # Feedbackversion is a numeric value, with 2 decimal numbers and 5 numbers in total. 
-    feedbackVersion  = db.Column(db.Numeric(5,2), unique=False, default=None)
+    feedbackVersion  = db.Column(db.Float, unique=False, default=None)
 
     def __repr__(self):
         return '<Scores {}>'.format(self.fileId)
